backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# OpenAI Integration
openai==1.6.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# PDDL & Planning
unified-planning==1.0.0

# Utilities
requests==2.31.0
Werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
